@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from 'react-dom/client';
 import "./index.css";
 
 function Square(props) {
@@ -111,8 +111,13 @@ class Game extends React.Component {
 
 // ====================================
 
-ReactDOM.render(<Game />, document.getElementById("root"));
-
+// ReactDOM.render(<Game />, document.getElementById("root"));
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <Game />
+  </React.StrictMode>
+);
 function calculateWinner(squares) {
   const lines = [
     [0, 1, 2],
